Fix for large files (It is the typical bug of many kioslave; the createUDSEntry function uses a "long int" as size instaed of the correct "long long int".)
svn path=/branches/KDE/3.5/kdesdk/kioslave/; revision=468882

+++ svn/svn.h
@@ -132,7 +132,7 @@ class kio_svnProtocol : public KIO::SlaveBase
 		};
 
 	private:
-		bool createUDSEntry( const QString& filename, const QString& user, long int size, bool isdir, time_t mtime, KIO::UDSEntry& entry);
+		bool createUDSEntry( const QString& filename, const QString& user, long long int size, bool isdir, time_t mtime, KIO::UDSEntry& entry);
 		apr_pool_t *pool;
 		int m_counter;
 };
